+++ src/ska_mid_dish_manager/devices/DishManagerDS.py
@@ -22,6 +22,10 @@ from tango import AttrWriteType, Database, DbDevInfo, DebugIt, DevFloat, DispLev
 from tango.server import attribute, command, device_property, run
 
 from ska_mid_dish_manager.component_managers.dish_manager_cm import DishManagerComponentManager
+from ska_mid_dish_manager.interface.input_validation import (
+    TrackLoadTableFormatting,
+    TrackTableTimestampError,
+)
 from ska_mid_dish_manager.models.dish_enums import (
     Band,
     CapabilityStates,
@@ -34,6 +38,10 @@ from ska_mid_dish_manager.models.dish_enums import (
 )
 
 DevVarLongStringArrayType = Tuple[List[ResultCode], List[Optional[str]]]
+
+# Used for input validation. Input samples to tracktable that is less that
+# TRACK_LOAD_FUTURE_THRESHOLD_SEC in the future are logged
+TRACK_LOAD_FUTURE_THRESHOLD_SEC = 5
 
 
 # pylint: disable=too-many-instance-attributes
@@ -51,9 +59,9 @@ class DishManager(SKAController):
     # -----------------
     # these values will be overwritten by values in
     # /charts/ska-mid-dish-manager/data in k8s deployment
-    DSDeviceFqdn = device_property(dtype=str, default_value="ska001/ds/manager")
-    SPFDeviceFqdn = device_property(dtype=str, default_value="ska001/spf/simulator")
-    SPFRxDeviceFqdn = device_property(dtype=str, default_value="ska001/spfrx/simulator")
+    DSDeviceFqdn = device_property(dtype=str, default_value="mid-dish/ds-manager/SKA001")
+    SPFDeviceFqdn = device_property(dtype=str, default_value="mid-dish/simulator-spfc/SKA001")
+    SPFRxDeviceFqdn = device_property(dtype=str, default_value="mid-dish/simulator-spfrx/SKA001")
     DishId = device_property(dtype=str, default_value="SKA001")
 
     def create_component_manager(self):
@@ -748,20 +756,23 @@ class DishManager(SKAController):
         # pylint: disable=attribute-defined-outside-init
         # Spectrum that is a multiple of 3 values:
         # - (timestamp, azimuth coordinate, elevation coordinate)
-        # i.e. [timestamp_0, az_pos_0, el_pos_0, ..., timestamp_n, az_pos_n, el_pos_n]
+        # i.e. [tai_0, az_pos_0, el_pos_0, ..., tai_n, az_pos_n, el_pos_n]
         self.logger.debug("programTrackTable write method called with table %s", table)
+
+        # perform input validation on table
+        try:
+            TrackLoadTableFormatting().check_track_table_input_valid(
+                table, TRACK_LOAD_FUTURE_THRESHOLD_SEC
+            )
+        except TrackTableTimestampError as te:
+            self.logger.error("TrackTableTimestampError: %s", te)
+        except ValueError as ve:
+            raise ve
+
         length_of_table = len(table)
-        if length_of_table > 0:
-            # Checks that the tables length is a multiple of 3
-            if length_of_table % 3 == 0:
-                sequence_length = length_of_table / 3
-                self.component_manager._track_load_table(sequence_length, table)
-                self._program_track_table = table
-            else:
-                raise ValueError(
-                    f"Length of table ({len(table)}) is not a multiple of 3 "
-                    "(timestamp, azimuth coordinate, elevation coordinate) as expected."
-                )
+        sequence_length = length_of_table / 3
+        self.component_manager._track_load_table(sequence_length, table)
+        self._program_track_table = table
 
     @attribute(
         dtype=int,