+++ tests/acceptance/test_stow.py
@@ -2,7 +2,7 @@
 import pytest
 import tango
 
-from ska_mid_dish_manager.models.dish_enums import DishMode, DSOperatingMode
+from ska_mid_dish_manager.models.dish_enums import DishMode
 
 
 # pylint: disable=too-many-locals,unused-argument
@@ -13,7 +13,7 @@ def test_stow_transition(
     monitor_tango_servers, event_store_class, dish_manager_proxy, ds_device_proxy
 ):
     """Test transition to STOW"""  # Get at least one device into a known state
-    ds_device_proxy.operatingMode = DSOperatingMode.STANDBY_FP
+    ds_device_proxy.SetStandbyFPMode()
 
     main_event_store = event_store_class()
     progress_event_store = event_store_class()